Refactor to use QuestionsRepository class
